requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.22
alembic==1.12.0
python-dotenv==1.0.0
passlib[bcrypt]
python-jose==3.3.0
psycopg2-binary==2.9.11
pydantic-settings